Refector of transaction handlers to a separate module
